requirements.txt
streamlit
python-dotenv
google-generativeai
sqlite-utils
mysql-connector-python
